+++ src/daemail/__main__.py
@@ -1,7 +1,6 @@
 from   codecs        import getdecoder
 from   collections   import namedtuple
 import locale
-import netrc
 import os
 import sys
 import traceback
@@ -10,65 +9,14 @@ import daemon
 from   daemon.daemon import DaemonError
 from   mailbits      import ContentType
 from   morecontext   import dirchanged
+from   outgoing      import from_config_file, get_default_configpath
 from   .             import __version__
 # Import runner instead of runner.CommandRunner etc. for mocking purposes
 from   .             import reporter, runner, senders
 from   .util         import AddressParamType, dt2stamp, dtnow, get_mime_type, \
                                 multiline822, show_argv
 
-DEFAULT_SENDMAIL = 'sendmail -i -t'
-
 outfile_type = click.Path(writable=True, dir_okay=False, resolve_path=True)
-
-def set_sender_cls(cls):
-    def callback(ctx, param, value):
-        if value is not None and value is not False:
-            ctx.params['sender_cls'] = cls
-            return value
-    return callback
-
-def use_smtp(ctx, param, value):
-    if value is not None \
-            and (ctx.params.get('sender_cls') is None
-                or not issubclass(ctx.params['sender_cls'],senders.SMTPSender)):
-        ctx.params['sender_cls'] = senders.SMTPSender
-    return value
-
-def pwd_getter(f):
-    def callback(ctx, param, value):
-        if value is not None and value is not False:
-            ctx.params['smtp_password_getter'] = f(value)
-    return callback
-
-def plain_pwd(password):
-    return lambda host, username: (username, password)
-
-def password_file(fp):
-    def get(host, username):
-        with fp:
-            smtp_password = fp.read()
-        # Remove no more than one line ending sequence:
-        if smtp_password.endswith('\n'):
-            smtp_password = smtp_password[:-1]
-        if smtp_password.endswith('\r'):
-            # This should be unreachable due to Python opening files with
-            # universal newlines enabled by default, but just in case...
-            smtp_password = smtp_password[:-1]  # pragma: no cover
-        return (username, smtp_password)
-    return get
-
-def netrc_getter(value):
-    def get(host, username):
-        nrc = netrc.netrc(None if value is True else value)
-        login = nrc.authenticators(host)
-        if login is not None:
-            if username is not None:
-                if login[0] in (None, '', username):
-                    return (username, login[2])
-            elif login[0] not in (None, ''):
-                return (login[0], login[2])
-        return (username, None)
-    return get
 
 def validate_encoding(ctx, param, value):
     if value is not None:
@@ -99,6 +47,13 @@ def get_cwd():
     __version__,
     '-V', '--version',
     message = '%(prog)s %(version)s',
+)
+@click.option(
+    "-c",
+    "--config",
+    type=click.Path(dir_okay=False),
+    default=get_default_configpath,
+    help="Specify the configuration file to use",
 )
 @click.option(
     '-C', '--chdir',
@@ -177,87 +132,11 @@ def get_cwd():
     help     = 'To: address of e-mail',
 )
 @click.option('-Z', '--utc', is_flag=True, help='Use UTC timestamps')
-@click.option(
-    '-s', '--sendmail',
-    metavar  = 'COMMAND',
-    callback = set_sender_cls(senders.CommandSender),
-    help     = 'Command for sending e-mail',
-)
-@click.option(
-    '--mbox',
-    help     = 'Append e-mail to this mbox file',
-    type     = outfile_type,
-    callback = set_sender_cls(senders.MboxSender),
-)
-@click.option(
-    '--smtp-host',
-    metavar  = 'HOST',
-    callback = use_smtp,
-    help     = 'SMTP server through which to send e-mail',
-)
-@click.option(
-    '--smtp-port',
-    type     = int,
-    metavar  = 'PORT',
-    callback = use_smtp,
-    help     = 'Connect to --smtp-host on this port',
-)
-@click.option(
-    '--smtp-username',
-    metavar  = 'USERNAME',
-    callback = use_smtp,
-    help     = 'Username for authenticating with --smtp-host',
-)
-@click.option(
-    '--smtp-password',
-    metavar      = 'PASSWORD',
-    callback     = pwd_getter(plain_pwd),
-    expose_value = False,
-    help         = 'Password for authenticating with --smtp-host',
-)
-@click.option(
-    '--smtp-password-file',
-    metavar      = 'FILE',
-    type         = click.File(),
-    callback     = pwd_getter(password_file),
-    expose_value = False,
-    help         = 'File containing password for --smtp-host',
-)
-@click.option(
-    '--netrc',
-    is_flag      = True,
-    callback     = pwd_getter(netrc_getter),
-    expose_value = False,
-    help         ='Fetch SMTP password from ~/.netrc file',
-)
-@click.option(
-    '--netrc-file',
-    type         = click.Path(dir_okay=False),
-    callback     = pwd_getter(netrc_getter),
-    expose_value = False,
-    help         = 'Fetch SMTP password from given netrc file',
-)
-# Implementing `--smtp-ssl` and `--smtp-starttls` as feature switches writing
-# to `sender_cls` won't work, as click will overwrite `sender_cls` with `None`
-# if neither option is given.
-@click.option(
-    '--smtp-ssl',
-    is_flag      = True,
-    callback     = set_sender_cls(senders.SMTP_SSLSender),
-    expose_value = False,
-    help         = 'Use SMTPS protocol',
-)
-@click.option(
-    '--smtp-starttls',
-    is_flag      = True,
-    callback     = set_sender_cls(senders.StartTLSSender),
-    expose_value = False,
-    help         = 'Use SMTP protocol with STARTTLS',
-)
 @click.argument('command')
 @click.argument('args', nargs=-1, type=click.UNPROCESSED)
 def main(
     command, args,
+    config,
     chdir, foreground,
     logfile,
     from_addr, to_addr,
@@ -265,30 +144,11 @@ def main(
     no_stdout, no_stderr, split,
     encoding, stderr_encoding, mime_type, stdout_filename,
     utc,
-    sendmail, mbox, dead_letter,
-    smtp_host, smtp_port, smtp_username,
-    sender_cls=None,
-    smtp_password_getter=None,
+    dead_letter,
 ):
     """ Daemonize a command and e-mail the results """
 
-    if sender_cls is None:
-        sender = senders.CommandSender(DEFAULT_SENDMAIL)
-    elif sender_cls is senders.CommandSender:
-        sender = sender_cls(sendmail)
-    elif sender_cls is senders.MboxSender:
-        sender = sender_cls(mbox)
-    else:
-        assert issubclass(sender_cls, senders.SMTPSender)
-        if smtp_host is None:
-            raise click.UsageError('--smtp-host is required for SMTP')
-        password = None
-        if smtp_password_getter is not None:
-            smtp_username, password = \
-                smtp_password_getter(smtp_host, smtp_username)
-        if smtp_username is not None and password is None:
-            password = click.prompt('SMTP password', hide_input=True, err=True)
-        sender = sender_cls(smtp_host, smtp_port, smtp_username, password)
+    sender = from_config_file(config, fallback=False)
 
     if encoding is None:
         encoding = locale.getpreferredencoding(True)
@@ -368,9 +228,7 @@ class Daemail(namedtuple('Daemail', 'runner reporter mailer')):
         s += '"To:" addresses:\n'
         for t in self.reporter.to_addrs:
             s += '  ' + str(t) + '\n'
-        s += 'Outgoing mail:\n'
-        for k,v in self.mailer.sender.about():
-            s += f'  {k}: {v}\n'
+        s += f'Outgoing mail class: {type(self.mailer.sender)}\n'
         s += 'Dead letter mbox: ' + repr(self.mailer.dead_letter_path) + '\n'
         s += 'Split stdout/stderr: ' + yesno(self.runner.split) + '\n'
         s += 'Capture stdout: ' + yesno(not self.runner.no_stdout) + '\n'